improved permalinks
adresses #6

Setup: 2 collections: pages, posts; 2 languages

* pages are found with expected urls
* pagination for sub pages has duplicated lang prefixes
* breadcrumbs for sub pages are disabled

I have a feeling, that permalinks fit very well in a setup with all entries in a single pages collection, but not in a setup with multiple collections for sub pages - at least not with the current application logic.

Using permalinks parallel to the current logic with slugs (and not as an alternative) should be easier to implement.